pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uptimemesh"
version = "0.1.0"
description = "UptimeMesh control-plane UI and API"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "sqlalchemy>=2.0",
  "greenlet>=3.0",
  "aiosqlite>=0.19",
  "pydantic-settings>=2.0",
  "alembic>=1.13",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
  "cryptography>=43.0",
]

[project.scripts]
uptimemesh = "app.cli:main"

[project.optional-dependencies]
dev = [
  "ruff>=0.3",
  "pytest>=8.0",
  "pyright>=1.1",
]

[tool.ruff]
line-length = 100

[tool.pyright]
pythonVersion = "3.11"
reportMissingTypeStubs = false

[tool.setuptools.packages.find]
include = ["app*", "migrations*"]
